requirements.txt
typing-extensions>=4.5.0,<4.9.0
apache-airflow==2.7.3
annotated-types>=0.4.0
pydantic>=2.5.0,<2.8.0
groq>=0.8.0,<0.11.0
numpy>=1.21.0,<1.25.0
pandas>=1.3.0,<2.1.0
scikit-learn>=1.0.0,<1.4.0
tenacity>=8.0.0
psutil>=5.8.0
kafka-python>=2.0.2
mysql-connector-python>=8.0.0
mlflow>=2.0.0
transformers>=4.20.0,<5.0.0
pyspark==3.5.5
mlflow 
prophet
pymysql
apache-airflow-providers-mysql